Only delete if modified

+++ data_updates/Python/iati_transactions.py
@@ -138,7 +138,8 @@ def main(args):
                 transaction_table = Table(DATA_TABLENAME, meta, schema=DATA_SCHEMA, autoload=True)
                 if_exists = "append"
         else:
-            conn.execute(transaction_table.delete().where(transaction_table.c.package_id == dataset["id"]))
+            if dataset["modified"] == True:
+                conn.execute(transaction_table.delete().where(transaction_table.c.package_id == dataset["id"]))
 
             flat_data.to_sql(name=TMP_DATA_TABLENAME, con=engine, schema=TMP_DATA_SCHEMA, index=False, if_exists=if_exists)
